chore(api): delete dead map-app skeleton from apps/api
apps/api's skeleton was copied from apps/map and still carried
page-app machinery that's dead weight for a pure oRPC API server:
admin/editor middleware guarding a nonexistent /admin route, a React
error boundary with no page tree to protect, browser/edge Sentry
config, and map-app leftovers in next.config.js. Also dedupes the
x-forwarded base-URL derivation into a shared getBaseUrl() util and
rewrites README.md to describe the actual service.

next-auth is kept as a dependency (still required by the
characterization parity-gate suite's session-cookie fixtures) despite
being flagged for removal in the original issue.

Part of the apps/api -> Hono migration epic (#644), Phase 1.

Closes #648
