--- apps/api/__tests__/setup.tsx
@@ -1,1 +1,0 @@
-import "@testing-library/jest-dom";
