--- apps/api/middleware/types.ts
@@ -1,3 +1,0 @@
-import type { NextProxy } from "next/server";
-
-export type MiddlewareFactory = (middleware: NextProxy) => NextProxy;
