+++ apps/api/next.config.js
@@ -34,27 +34,8 @@ const config = {
     "/*": ["../../node_modules/.pnpm/@img+sharp-libvips-*/**/*"],
   },
 
-  images: {
-    remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "imgur.com",
-        pathname: "/*",
-      },
-    ],
-  },
-
   /** We already do typechecking as a separate task in CI */
   typescript: { ignoreBuildErrors: true },
-  redirects: async () => {
-    return [
-      {
-        source: "/map",
-        destination: "/",
-        permanent: true,
-      },
-    ];
-  },
 };
 
 export default withSentryConfig(config, {
@@ -72,10 +53,4 @@ export default withSentryConfig(config, {
 
   // Upload a larger set of source maps for prettier stack traces (increases build time)
   widenClientFileUpload: true,
-
-  // Route browser requests to Sentry through a Next.js rewrite to circumvent ad-blockers.
-  // This can increase your server load as well as your hosting bill.
-  // Note: Check that the configured route will not match with your Next.js middleware, otherwise reporting of client-
-  // side errors will fail.
-  tunnelRoute: "/monitoring",
 });
