+++ apps/api/package.json
@@ -41,17 +41,13 @@
     "@acme/tsconfig": "workspace:*",
     "@acme/vitest-config": "workspace:*",
     "@orpc/client": "catalog:",
-    "@testing-library/jest-dom": "catalog:",
-    "@testing-library/react": "catalog:",
     "@types/node": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
-    "@vitejs/plugin-react": "catalog:",
     "@vitest/coverage-v8": "catalog:",
     "eslint": "catalog:",
     "jiti": "catalog:",
     "jose": "catalog:",
-    "jsdom": "catalog:",
     "prettier": "catalog:",
     "typescript": "catalog:",
     "vite": "catalog:",
