--- apps/api/sentry.edge.config.ts
@@ -1,32 +1,0 @@
-// This file configures the initialization of Sentry for edge features (middleware, edge routes, and so on).
-// The config you add here will be used whenever one of the edge features is loaded.
-// This config is also required when running locally.
-// https://docs.sentry.io/platforms/javascript/guides/nextjs/
-import * as Sentry from "@sentry/nextjs";
-import { captureConsoleIntegration } from "@sentry/nextjs";
-
-import { env } from "~/env";
-
-if (env.NODE_ENV === "production") {
-  const channel = env.NEXT_PUBLIC_CHANNEL;
-  Sentry.init({
-    dsn: "https://7174fea65c117ea4b71977da953bb4d9@o4509266839797760.ingest.us.sentry.io/4509270283714560",
-
-    // Define how likely traces are sampled. Adjust this value in production, or use tracesSampler for greater control.
-    tracesSampleRate: 1,
-
-    // Setting this option to true will print useful information to the console while you're setting up Sentry.
-    debug: false,
-
-    environment:
-      channel === "prod"
-        ? "production"
-        : channel === "staging"
-          ? "staging"
-          : "development",
-
-    // Capture console.error calls and send them to Sentry
-    // This ensures caught errors that are logged still get reported
-    integrations: [captureConsoleIntegration({ levels: ["error"] })],
-  });
-}
