--- apps/api/src/app/global-error.tsx
@@ -1,26 +1,0 @@
-"use client";
-
-import { useEffect } from "react";
-import * as Sentry from "@sentry/nextjs";
-
-interface GlobalErrorProps {
-  error: Error & { digest?: string };
-  reset: () => void;
-}
-
-export default function GlobalError({ error, reset }: GlobalErrorProps) {
-  useEffect(() => {
-    Sentry.captureException(error);
-  }, [error]);
-
-  return (
-    <html lang="en">
-      <body>
-        <h2>Something went wrong.</h2>
-        <button type="button" onClick={() => reset()}>
-          Try again
-        </button>
-      </body>
-    </html>
-  );
-}
