+++ apps/api/src/instrumentation.ts
@@ -4,10 +4,6 @@ export async function register() {
   if (process.env.NEXT_RUNTIME === "nodejs") {
     await import("../sentry.server.config");
   }
-
-  if (process.env.NEXT_RUNTIME === "edge") {
-    await import("../sentry.edge.config");
-  }
 }
 
 export const onRequestError = Sentry.captureRequestError;
