+++ apps/api/vitest.config.ts
@@ -1,15 +1,12 @@
 import { coverageExclude, coverageInclude } from "@acme/vitest-config";
-import react from "@vitejs/plugin-react";
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
-  plugins: [react()],
   resolve: { tsconfigPaths: true },
   test: {
     globals: true,
-    environment: "jsdom",
+    environment: "node",
     env: { NODE_ENV: "test" },
-    setupFiles: ["__tests__/setup.tsx"],
     coverage: {
       provider: "v8",
       reporter: ["text", "json", "html"],
@@ -18,10 +15,10 @@ export default defineConfig({
       exclude: coverageExclude,
       thresholds: {
         autoUpdate: true,
-        statements: 91.04,
-        branches: 93.47,
-        functions: 80,
-        lines: 91.04,
+        statements: 98.18,
+        branches: 100,
+        functions: 87.5,
+        lines: 98.18,
       },
     },
     exclude: [
@@ -31,10 +28,5 @@ export default defineConfig({
       "**/cypress/**",
       "**/.{idea,git,cache,output,temp}/**",
     ],
-    server: {
-      deps: {
-        inline: ["vitest-canvas-mock", "jest-canvas-mock"],
-      },
-    },
   },
 });
